requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.5
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0